chore: move Signature field support to \Extensions
