+++ src/Extensions/GFSignature/Type/Enum/SignatureFieldBorderWidthEnum.php
@@ -2,11 +2,13 @@
 /**
  * Enum Type - SignatureFieldBorderWidthEnum
  *
- * @package WPGraphQL\GF\Type\Enum,
+ * @package WPGraphQL\GF\Extensions\GFSignature\Type\Enum,
  * @since   0.4.0
  */
 
-namespace WPGraphQL\GF\Type\Enum;
+namespace WPGraphQL\GF\Extensions\GFSignature\Type\Enum;
+
+use WPGraphQL\GF\Type\Enum\AbstractEnum;
 
 /**
  * Class - SignatureFieldBorderWidthEnum
